Fixed compilation error and warnings

+++ Scheduler/Source/MTFiberContext.cpp
@@ -1,17 +1,17 @@
 // The MIT License (MIT)
-// 
+//
 // 	Copyright (c) 2015 Sergey Makeev, Vadim Slyusarev
-// 
+//
 // 	Permission is hereby granted, free of charge, to any person obtaining a copy
 // 	of this software and associated documentation files (the "Software"), to deal
 // 	in the Software without restriction, including without limitation the rights
 // 	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 // 	copies of the Software, and to permit persons to whom the Software is
 // 	furnished to do so, subject to the following conditions:
-// 
+//
 //  The above copyright notice and this permission notice shall be included in
 // 	all copies or substantial portions of the Software.
-// 
+//
 // 	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // 	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // 	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -75,7 +75,7 @@ namespace MT
 		MT_ASSERT(threadContext->thread.IsCurrentThread(), "Thread context sanity check failed");
 
 		MT_VERIFY(group != currentGroup, "Can't wait the same group. Deadlock detected!", return);
-		MT_VERIFY(group != MT::INVALID_GROUP && group >= 0 && group < MT_MAX_GROUPS_COUNT, "Invalid group!", return);
+		MT_VERIFY(group != MT::INVALID_GROUP && group >= (int)0 && group < (int)MT_MAX_GROUPS_COUNT, "Invalid group!", return);
 
 		TaskScheduler::TaskGroupDescription  & groupDesc = threadContext->taskScheduler->GetGroupDesc(group);
